+++ docs/releasenotes.rst
@@ -15,6 +15,13 @@ Next release
 * The `mpi_implementations` configuration option is no longer needed and has
   been deprecated.
 
+* MyQueue no longer tries to keep track of all your ``.myqueue/`` folders.
+  Consequently, the ``--all`` option has been removed from the :ref:`list`,
+  :ref:`kick` and :ref:`sync` commands.
+
+* There is now one background daemon per ``.myqueue/`` folder.  See
+  :ref:`daemon process`.
+
 
 Version 21.7.0
 ==============